requirements.txt
torch==2.5.1
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
numpy==2.2.1
coverage==7.6.10
pytest==8.3.4
lightning==2.5.0
lightning[extra]==2.5.0
torchvision==0.20.1
seaborn==0.13.2
python-multipart==0.0.20
ruff==0.9.0
dill==0.3.9
google-cloud-storage==2.19.0
streamlit==1.41.1
requests==2.32.3
pillow==11.1.0
httpx==0.24.1
loguru==0.7.3
python-dotenv==1.0.1
wandb==0.19.4
invoke==2.2.0
locust==2.32.6
scikit-learn==1.6.1
albumentations==2.0.1
